requirements.txt
mysql-connector==2.1.6
PyQt5==5.11.2
PyQt5-sip==4.19.12
setuptools==28.8.0